WMSRequestHandler.java getServiceProvider methode aangepast en functionele volgorde verbeterd. Gehele project teruggezet naar Java versie 1.4. Try-catches verwijderd daar waar mogelijk en vervangen door throws. System.out.println door het gehele project verwijderd. Foutmeldingen die naar de serverpagina gestuurd worden aangepast. KBConstants aan ServiceProvider en ContactInformation toegevoegd

+++ src/nl/b3p/kaartenbalie/core/server/Identifier.java
@@ -65,7 +65,7 @@ public class Identifier implements XMLElement {
     public Object clone() {
         Identifier cloneIdent           = new Identifier();
         if (null != this.id) {
-            cloneIdent.id               = new Integer(this.id);
+            cloneIdent.id               = new Integer(this.id.intValue());
         }
         if (null != this.authorityName) {
             cloneIdent.authorityName    = new String(this.authorityName);
